requirements.txt
pymupdf
pandas
python-dotenv
tqdm
sentence-transformers
faiss-cpu
botocore
streamlit